+++ src/xmpp/biboumi_adhoc_commands.hpp
@@ -20,7 +20,7 @@ void ConfigureIrcServerStep2(XmppComponent&, AdhocSession& session, XmlNode& com
 void ConfigureIrcChannelStep1(XmppComponent&, AdhocSession& session, XmlNode& command_node);
 void insert_irc_channel_configuration_form(XmlNode& node, const Jid& requester, const Jid& target);
 void ConfigureIrcChannelStep2(XmppComponent&, AdhocSession& session, XmlNode& command_node);
-bool handle_irc_channel_configuration_form(const XmlNode& node, const Jid& requester, const Jid& target);
+bool handle_irc_channel_configuration_form(XmppComponent&, const XmlNode& node, const Jid& requester, const Jid& target);
 
 void DisconnectUserFromServerStep1(XmppComponent&, AdhocSession& session, XmlNode& command_node);
 void DisconnectUserFromServerStep2(XmppComponent&, AdhocSession& session, XmlNode& command_node);